ENH: finished masking subsection draft ready for review
