+++ thesis.fdb_latexmk
@@ -1,29 +1,34 @@
 # Fdb version 3
-["bibtex thesis"] 1463161421 "thesis.aux" "thesis.bbl" "thesis" 
-  "bmk/bmk.aux" 1463161422 2443 f54a016978080f00985feb4beb8d20d2 ""
-  "misc/intro.aux" 1463161421 1749 fc18368a4040e236e91c6eb55b0c7fbd ""
-  "pdf/pdf.aux" 1463161422 10074 44644288b547bd4006d4fc10df8f6b58 ""
-  "pes_e/pes_e.aux" 1463161421 3511 886dceabffd004d39291d9dee329e339 ""
-  "thesis.aux" 1463161422 3927 81fc69247bca84b69d44b1edf697776f ""
+["bibtex thesis"] 1463177190 "thesis.aux" "thesis.bbl" "thesis" 
+  "bmk/bmk.aux" 1463177191 2443 5c5379c44aba5ad1c92edbe221eabc5b ""
+  "misc/intro.aux" 1463177190 1749 bcd964aa83d0d16c9acc4098a95583b2 ""
+  "pdf/pdf.aux" 1463177191 11773 4fc5c89d574a96d0a26da6b28ded27f8 ""
+  "pes_e/pes_e.aux" 1463177190 3511 22c4d0efe779ff19e56a86702c50df58 ""
+  "thesis.aux" 1463177191 3927 e6a7f7e0387c9fb1131cc2fa0fd79818 ""
   "thesis.bcf" 1462791291 11850 713f0ce63cb6607c15856f34921159ef ""
   (generated)
+  "thesis.blg"
   "thesis.bbl"
-["pdflatex"] 1463161421 "thesis.tex" "thesis.pdf" "thesis" 
-  "./bmk/bmk.aux" 1463161422 2443 f54a016978080f00985feb4beb8d20d2 ""
+["pdflatex"] 1463177190 "thesis.tex" "thesis.pdf" "thesis" 
+  "./bmk/bmk.aux" 1463177191 2443 5c5379c44aba5ad1c92edbe221eabc5b ""
   "./bmk/bmk.tex" 1463077044 232 bc421e2a2b6766246780cf110bd1130e ""
   "./misc/abstract.tex" 1462550623 1 68b329da9893e34099c7d8ad5cb9c940 ""
   "./misc/dedication.tex" 1462550623 1 68b329da9893e34099c7d8ad5cb9c940 ""
-  "./misc/intro.aux" 1463161421 1749 fc18368a4040e236e91c6eb55b0c7fbd ""
+  "./misc/intro.aux" 1463177190 1749 bcd964aa83d0d16c9acc4098a95583b2 ""
   "./misc/intro.tex" 1462551336 64 812990139116f6fa00a8c434234c3e8f ""
   "./misc/thanks.tex" 1462550623 1 68b329da9893e34099c7d8ad5cb9c940 ""
   "./pdf/figures/bad_bin_dead_pixel_masked_100.png" 1463156961 317693 1d8378a2cd86ae23a7232a1ee7e78369 ""
   "./pdf/figures/bad_bin_dead_pixel_masked_1000.png" 1463157266 490116 6cfc2413bb8d214d28009a37d8ba41e6 ""
   "./pdf/figures/bad_bin_dead_pixel_masked_300.png" 1463157065 368521 6f338f5b9555c71433c18a07bf36dad9 ""
   "./pdf/figures/bad_bin_dead_pixel_masked_500.png" 1463157168 411702 2a6cff92abd6e6ffd8ad7c15dfe3057b ""
+  "./pdf/figures/combined_masked.png" 1463176866 21537 f729e874ee6a95f16c705e7cd5a0f0d7 ""
+  "./pdf/figures/combined_raw.png" 1463176863 93038 c105e6f611818d589a20ce8c05333b80 ""
   "./pdf/figures/dead_pixel_masked_100.pdf" 1463155678 1298002 7b91d8a35f178ccee446c0c41f12e65d ""
   "./pdf/figures/dead_pixel_masked_1000.pdf" 1463155763 1579883 12ea186e8c052e05f74e17b2eb92ca0d ""
   "./pdf/figures/dead_pixel_masked_300.pdf" 1463155704 1361987 9bf16d05984cf0ca8c0bc411943c2b56 ""
   "./pdf/figures/dead_pixel_masked_500.pdf" 1463155732 1426683 bef05949067d347b111be5ff46dab5f4 ""
+  "./pdf/figures/exp_masked.png" 1463175513 12852 0c5b1df4acc18f9d44f11a40d0f0b779 ""
+  "./pdf/figures/exp_raw.png" 1463175510 91849 e502418d92a25854b7a363f6c1565a6d ""
   "./pdf/figures/masked_10.pdf" 1463157903 1274950 2c515274fa57ab93cdf681267a07b338 ""
   "./pdf/figures/masked_30.pdf" 1463157933 1275792 712e7915aebfb7c2c5a26999dfc55186 ""
   "./pdf/figures/masked_50.pdf" 1463157964 1276212 e816a32e282b6c599df8d21f4b8c1cad ""
@@ -40,10 +45,12 @@
   "./pdf/figures/rotate_masked_50.png" 1463158438 218944 be7d09b600366787f8dd5636e247495c ""
   "./pdf/figures/rotate_missed_50.png" 1463158442 11419 8658e19a8affe0d1d50ce78062766b94 ""
   "./pdf/figures/rotate_raw_50.png" 1463158434 233949 95534614b477cac9fc411954944e3f5e ""
-  "./pdf/mask.tex" 1463159969 7289 a739c54b0e21a4cab3b8a803675e2f54 ""
-  "./pdf/pdf.aux" 1463161422 10074 44644288b547bd4006d4fc10df8f6b58 ""
+  "./pdf/figures/single_xtal_masked.png" 1463176550 57057 4b1db89834cf4103955472aed35965d3 ""
+  "./pdf/figures/single_xtal_raw.png" 1463176547 55096 14764146e916446991fe9283f663082e ""
+  "./pdf/mask.tex" 1463177188 9436 e59cf23c4cb1a84d015c9c22974bd45d ""
+  "./pdf/pdf.aux" 1463177191 11773 4fc5c89d574a96d0a26da6b28ded27f8 ""
   "./pdf/pdf.tex" 1463158727 10032 5ee6eb366bf48cb0ee8f82c1eaf71ccc ""
-  "./pes_e/pes_e.aux" 1463161421 3511 886dceabffd004d39291d9dee329e339 ""
+  "./pes_e/pes_e.aux" 1463177190 3511 22c4d0efe779ff19e56a86702c50df58 ""
   "./pes_e/pes_e.tex" 1463078060 12669 6b8fd20fcf9479a76725fa00fafe1c2a ""
   "/usr/share/texlive/texmf-dist/tex/context/base/supp-pdf.mkii" 1337017135 71627 94eb9990bed73c364d7f53f960cc8c5b ""
   "/usr/share/texlive/texmf-dist/tex/generic/ifxetex/ifxetex.sty" 1284331290 1458 43ab4710dc82f3edeabecd0d099626b2 ""
@@ -167,17 +174,17 @@
   "/usr/share/texmf/tex/latex/pgf/utilities/pgfrcs.sty" 1302059272 325 2bcd023400636339210573e2b3ee298b ""
   "/usr/share/texmf/tex/latex/xcolor/xcolor.sty" 1339077425 55224 a43bab84e0ac5e6efcaf9a98bde73a94 ""
   "biblatex.sty" 1462552210 3393 b10054953b24eb42db30241eb250cd6a ""
-  "thesis.aux" 1463161422 3927 81fc69247bca84b69d44b1edf697776f ""
-  "thesis.bbl" 1463161421 0 d41d8cd98f00b204e9800998ecf8427e "bibtex thesis"
-  "thesis.lof" 1463161422 3874 d8ee1146bf5d668a0da032c7b051febb ""
-  "thesis.lot" 1463161422 180 dd876473cdbf3ae298ff9bc82038ceef ""
+  "thesis.aux" 1463177191 3927 e6a7f7e0387c9fb1131cc2fa0fd79818 ""
+  "thesis.bbl" 1463177190 0 d41d8cd98f00b204e9800998ecf8427e "bibtex thesis"
+  "thesis.lof" 1463177191 4758 dba8dd751a5069961a05ea13dbf9290a ""
+  "thesis.lot" 1463177191 180 dd876473cdbf3ae298ff9bc82038ceef ""
   "thesis.tex" 1463161420 11658 108c7b16820eab06323186d921b6cb43 ""
-  "thesis.toc" 1463161422 4961 ddfb68684c4af61eb86db554dedb0c76 ""
+  "thesis.toc" 1463177191 4961 4360b7ce38a4e3011afae2187d0ecd0c ""
   "uscthesis.cls" 1462550623 34933 468fb32e999bd20ab581757aa4de4579 ""
   (generated)
   "thesis.log"
   "thesis.toc"
   "thesis.lof"
   "thesis.pdf"
-  "thesis.aux"
   "thesis.lot"
+  "thesis.aux"
